requirements.txt
# Core dependencies
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==0.21.1
pymongo==4.3.3
requests==2.28.2
beautifulsoup4==4.11.2
python-multipart==0.0.6
pydantic==1.10.7
aiofiles==0.8.0

# File processing dependencies
PyPDF2==3.0.1
python-docx==0.8.11

# Testing dependencies
pytest==7.3.0
mongomock==4.1.2
httpx==0.24.0